ACM/Tests/ACM.BLTest/ProductRepositoryTest.cs
using ACM.BL;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace ACM.BLTest
{
    [TestClass]
    public class ProductRepositoryTest
    {
        [TestMethod]
        public void RetrieveTest()
        {
            //-- Arrange
            var ProductRepository = new ProductRepository();
            var expected = new Product(2)
            {
                ProductName = "Flowers",
                ProductDescription = "Yellow sunflowers",
                CurrentPrice = 39.99M
            };

            //-- Act
            var actual = ProductRepository.Retrieve(2);

            //-- Assert
            Assert.AreEqual(expected.ProductName, actual.ProductName);
            Assert.AreEqual(expected.ProductDescription, actual.ProductDescription);
            Assert.AreEqual(expected.CurrentPrice, actual.CurrentPrice);
        }
    }
}
